+++ IFLYADLibSimple/IFLYADLibSimple/biz/splash/IFLYSplashViewController.m
@@ -138,7 +138,7 @@
 
 - (void)destroyAd {
     [self destroyAdSilently];
-    [self updateStatus:@"已销毁" color:UIColor.systemTealColor];
+    [self updateStatus:@"已销毁" color:[IFLYADUtil demoTealColor]];
     [self log:@"Destroy"];
 }
 
@@ -180,7 +180,7 @@
                                       ad.hasVideoTemplate ? @"YES" : @"NO",
                                       ad.isLandscapeTemplate ? @"YES" : @"NO",
                                       [ad ecpm]]];
-    [self updateStatus:@"已加载，等待素材 ready" color:UIColor.systemIndigoColor];
+    [self updateStatus:@"已加载，等待素材 ready" color:[IFLYADUtil demoIndigoColor]];
 }
 
 - (void)splashAdDidReady:(IFLYSplashAd *)ad {
@@ -203,7 +203,7 @@
 
 - (void)splashAdDidClose:(IFLYSplashAd *)ad {
     [self log:@"splashAdDidClose"];
-    [self updateStatus:@"开屏已关闭" color:UIColor.systemTealColor];
+    [self updateStatus:@"开屏已关闭" color:[IFLYADUtil demoTealColor]];
 }
 
 - (void)splashAdDidSkip:(IFLYSplashAd *)ad {